+++ src/lib/services/publicProjectService.ts
@@ -1,15 +1,17 @@
 import { useWallet } from "@cosmos-kit/react";
+import type { UseQueryResult } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
 import axios from "axios";
 import { useCallback } from "react";
 
 import { CELATONE_API_ENDPOINT, getChainApiPath, getMainnetApiPath } from "env";
 import type {
+  Contract,
   Option,
+  PublicInfo,
+  PublicProjectInfo,
   RawContract,
   RawPublicProjectInfo,
-  PublicProjectInfo,
-  Contract,
 } from "lib/types";
 
 const parseContract = (raw: RawContract): Contract => ({
@@ -19,10 +21,11 @@ const parseContract = (raw: RawContract): Contract => ({
   slug: raw.slug,
 });
 
-export const usePublicProjectsQuery = () => {
+export const usePublicProjects = () => {
   const { currentChainRecord } = useWallet();
 
   const queryFn = useCallback(async () => {
+    // eslint-disable-next-line sonarjs/no-duplicate-string
     if (!currentChainRecord) throw new Error("No chain selected");
 
     return axios
@@ -44,7 +47,7 @@ export const usePublicProjectsQuery = () => {
   });
 };
 
-export const usePublicProjectBySlugQuery = (slug: Option<string>) => {
+export const usePublicProjectBySlug = (slug: Option<string>) => {
   const { currentChainRecord } = useWallet();
   const queryFn = useCallback(async (): Promise<Option<PublicProjectInfo>> => {
     if (!slug) throw new Error("No project selected");
@@ -64,5 +67,27 @@ export const usePublicProjectBySlugQuery = (slug: Option<string>) => {
   return useQuery(["public_project_by_slug"], queryFn, {
     keepPreviousData: true,
     enabled: !!slug,
+  });
+};
+
+export const usePublicProjectByContractAddress = (
+  contractAddress: Option<string>
+): UseQueryResult<PublicInfo> => {
+  const { currentChainRecord } = useWallet();
+  const queryFn = useCallback(async () => {
+    if (!contractAddress) throw new Error("Contract address not found");
+    if (!currentChainRecord) throw new Error("No chain selected");
+    return axios
+      .get<PublicInfo>(
+        `${CELATONE_API_ENDPOINT}/contracts/${getChainApiPath(
+          currentChainRecord.chain.chain_name
+        )}/${currentChainRecord.chain.chain_id}/${contractAddress}`
+      )
+      .then(({ data: projectInfo }) => projectInfo);
+  }, [contractAddress, currentChainRecord]);
+
+  return useQuery(["public_project_by_contract_address"], queryFn, {
+    keepPreviousData: true,
+    enabled: !!contractAddress,
   });
 };